Move Boot/ and board/ into a single unified BSP directory
