+++ BoardSupport/qemu-aarch64/Boot/CMakeLists.txt
@@ -1,9 +1,3 @@
-LIST(APPEND BOOT_HEADERS DeviceTree/DeviceTree.hpp)
-LIST(APPEND BOOT_SOURCES DeviceTree/DeviceTree.cpp)
-
-LIST(APPEND BOOT_HEADERS PSCI/PSCI.hpp)
-LIST(APPEND BOOT_SOURCES PSCI/PSCI.cpp)
-
 LIST(APPEND BOOT_HEADERS BootDriver.hpp)
 LIST(APPEND BOOT_SOURCES BootDriver.cpp)
