+++ CMakeLists.txt
@@ -28,6 +28,8 @@ INCLUDE_DIRECTORIES(${CMAKE_SOURCE_DIR})
 INCLUDE_DIRECTORIES(${CMAKE_BINARY_DIR})
 INCLUDE_DIRECTORIES(Common)
 INCLUDE_DIRECTORIES(System)
+INCLUDE_DIRECTORIES(${CMAKE_SOURCE_DIR}/arch/${CMAKE_SYSTEM_PROCESSOR})
+INCLUDE_DIRECTORIES(BoardSupport/Common)
 
 
 # Get definitions specific to this board, or the mock board for tests
@@ -36,16 +38,14 @@ IF (${ONLY_TESTS})
 	FIND_PACKAGE(Threads REQUIRED)
 	FIND_PACKAGE(GTest CONFIG)
 	
-	INCLUDE_DIRECTORIES(board/mock)
+	INCLUDE_DIRECTORIES(BoardSupport/mock)
 	INCLUDE_DIRECTORIES(${GTEST_INCLUDE_DIRS})
 	
-	ADD_SUBDIRECTORY(board/mock)
+	ADD_SUBDIRECTORY(BoardSupport/mock)
 	ADD_SUBDIRECTORY(tests)
 ELSE()
 	ADD_SUBDIRECTORY(${TARGET_SYSTEM_BOARD})
 ENDIF()
-
-INCLUDE_DIRECTORIES(${CMAKE_SOURCE_DIR}/arch/${CMAKE_SYSTEM_PROCESSOR})
 
 SET(CMAKE_INCLUDE_CURRENT_DIR ON)
 
@@ -55,7 +55,6 @@ ADD_SUBDIRECTORY(arch)
 ADD_SUBDIRECTORY(Common)
 ADD_SUBDIRECTORY(Devices)
 ADD_SUBDIRECTORY(System)
-ADD_SUBDIRECTORY(Boot)
 
 # Also go and add in some external dependencies...
 INCLUDE_DIRECTORIES(externals/printf)
