--- board/qemu-aarch64/CMakeLists.txt
@@ -1,6 +1,0 @@
-LIST(APPEND BOARD_HEADERS Board.hpp)
-LIST(APPEND BOARD_SOURCES Board.cpp)
-
-ADD_LIBRARY(Board STATIC
-	${BOARD_HEADERS}
-	${BOARD_SOURCES})
