+++ cmake/qemu-aarch64.cmake
@@ -9,7 +9,7 @@ SET(CMAKE_CXX_COMPILER "clang++")
 SET(CMAKE_ASM_COMPILER "aarch64-none-elf-as")
 SET(CMAKE_LINKER "ld.lld")
 SET(CMAKE_OBJCOPY "aarch64-none-elf-objcopy")
-SET(LINKER_SCRIPT ${CMAKE_SOURCE_DIR}/board/qemu-aarch64/linkin.ld)
+SET(LINKER_SCRIPT ${CMAKE_SOURCE_DIR}/BoardSupport/qemu-aarch64/linkin.ld)
 SET(CMAKE_ASM_NASM_OBJECT_FORMAT "elf64")
 SET(CMAKE_C_FLAGS "-target aarch64-none-elf -mgeneral-regs-only -mcmodel=large -ffreestanding -nostdlib -c -Wall -Wextra -Wvla -mcpu=cortex-a72 -mtune=cortex-a72 -fno-builtin ${CMAKE_C_FLAGS}")
 SET(CMAKE_C_LINK_EXECUTABLE "<CMAKE_LINKER> -m aarch64elf -nostdlib <OBJECTS> -o <TARGET> <LINK_LIBRARIES>")
@@ -24,7 +24,7 @@ SET(FINAL_CMAKE_CXX_LINK_EXECUTABLE "<CMAKE_LINKER> -m aarch64elf -nostdlib -T $
 
 SET(TARGET_SYSTEM "qemu-aarch64-virt")
 SET(TARGET_SYSTEM_QEMU_AARCH64_VIRT ON)
-SET(TARGET_SYSTEM_BOARD "board/qemu-aarch64")
+SET(TARGET_SYSTEM_BOARD "BoardSupport/qemu-aarch64")
 
 # Be sure about the name...
 SET(TARGET_PROCESSOR "aarch64")
